+++ laspy/laswriter.py
@@ -72,8 +72,6 @@ class LasWriter:
         self.encoding_errors = encoding_errors
         self.header = deepcopy(header)
         self.header.partial_reset()
-        self.header.maxs = [np.finfo("f8").min] * 3
-        self.header.mins = [np.finfo("f8").max] * 3
 
         self.dest = dest
         self.done = False
@@ -132,7 +130,7 @@ class LasWriter:
         if points.point_format != self.header.point_format:
             raise LaspyException("Incompatible point formats")
 
-        self.header.update(points)
+        self.header.grow(points)
         self.point_writer.write_points(points)
 
     def write_evlrs(self, evlrs: VLRList) -> None:
@@ -169,6 +167,11 @@ class LasWriter:
         if self.point_writer is not None:
             if not self.done:
                 self.point_writer.done()
+
+            if self.header.point_count == 0:
+                self.header.maxs = [0.0, 0.0, 0.0]
+                self.header.mins = [0.0, 0.0, 0.0]
+
             self.point_writer.write_updated_header(self.header, self.encoding_errors)
         if self.closefd:
             self.dest.close()